(supplies) added more dto's (products) added abstractions on each application layer (entities, repositories, services, controllers) and tests, added more dto's

+++ supplies/src/main/java/rtuitlab/supplies/services/AbstractService.java
@@ -1,8 +1,8 @@
 package rtuitlab.supplies.services;
 
 import lombok.AllArgsConstructor;
-import rtuitlab.supplies.dto.AbstractGetDTO;
-import rtuitlab.supplies.dto.AbstractPostPutDTO;
+import rtuitlab.supplies.dto.*;
+import rtuitlab.supplies.exceptions.EntityNotFoundException;
 import rtuitlab.supplies.mappers.CommonMapper;
 import rtuitlab.supplies.models.AbstractDocument;
 import rtuitlab.supplies.repositories.CommonRepository;
@@ -11,43 +11,47 @@ import java.util.List;
 import java.util.stream.Collectors;
 
 @AllArgsConstructor
-public class AbstractService<E extends AbstractDocument, R extends CommonRepository<E>, G extends AbstractGetDTO, P extends AbstractPostPutDTO, M extends CommonMapper<E, G, P>> implements CommonService<E, G, P> {
+public class AbstractService<E extends AbstractDocument, R extends CommonRepository<E>, Get extends AbstractGetDTO, Post extends AbstractPostDTO, Put extends AbstractPutDTO, Posted extends AbstractPostedDTO, Updated extends AbstractUpdatedDTO, M extends CommonMapper<E, Get, Post, Put, Posted, Updated>> implements CommonService<E, Get, Post, Put, Posted, Updated> {
 
     protected final R repository;
     protected final M mapper;
     @Override
-    public List<G> getAll() {
-        return repository.findAll().stream().map(mapper::entityToDTO).collect(Collectors.toList());
+    public List<Get> getAll() {
+        return repository.findAll().stream().map(mapper::entityToGetDTO).collect(Collectors.toList());
     }
 
     @Override
-    public G getById(String id) {
-        return mapper.entityToDTO(repository.findById(id).orElse(null));
+    public Get getById(String id) throws EntityNotFoundException {
+        return mapper.entityToGetDTO(repository.findById(id).orElseThrow(() -> new EntityNotFoundException(id)));
     }
 
     @Override
-    public List<G> create(P p) {
-        repository.save(mapper.postPutDTOToEntity(p));
-        return repository.findAll().stream().map(mapper::entityToDTO).collect(Collectors.toList());
+    public List<Posted> create(Post p) {
+        repository.save(mapper.postDTOToEntity(p));
+        return repository.findAll().stream().map(mapper::entityToPostedDTO).collect(Collectors.toList());
     }
 
     @Override
-    public G update(String id, P p) {
-        E e = mapper.postPutDTOToEntity(p);
+    public Updated update(String id, Put p) throws EntityNotFoundException {
+        if(!repository.existsById(id))
+            throw new EntityNotFoundException(id);
+        E e = mapper.putDTOToEntity(p);
         e.setId(id);
         repository.save(e);
-        return mapper.entityToDTO(repository.findById(id).orElse(null));
+        return mapper.entityToUpdatedDTO(repository.findById(id).orElse(null));
     }
 
     @Override
-    public List<G> deleteById(String id) {
+    public List<Get> deleteById(String id) throws EntityNotFoundException {
+        if(!repository.existsById(id))
+            throw new EntityNotFoundException(id);
         repository.deleteById(id);
-        return repository.findAll().stream().map(mapper::entityToDTO).collect(Collectors.toList());
+        return repository.findAll().stream().map(mapper::entityToGetDTO).collect(Collectors.toList());
     }
 
     @Override
-    public List<G> deleteAll() {
+    public List<Get> deleteAll() {
         repository.deleteAll();
-        return repository.findAll().stream().map(mapper::entityToDTO).collect(Collectors.toList());
+        return repository.findAll().stream().map(mapper::entityToGetDTO).collect(Collectors.toList());
     }
 }
